+++ scrapyd/exceptions.py
@@ -4,3 +4,7 @@ class ScrapydError(Exception):
 
 class MissingRequiredArgument(ScrapydError):
     """Raised if a required argument is missing"""
+
+
+class RunnerError(ScrapydError):
+    """Raised if the runner returns an error code"""